pyproject.toml
[project]
name = "task-printer"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-escpos[usb]", "imgkit"]

[dependency-groups]
dev = ["pytest>=8.4.1"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
